[PATCH] kfree list nodes after deleting them from free_client_list in batgat
list_del will only move the next pointer of the previous entry to the next entry
and the do the same with the previous pointer of the next entry. The entry
itself will not be touched.
We allocated new memory for the list entry so we must free them or otherwise we
would only leak memory and run out of it after enough connecting and leaving
clients.

Signed-off-by: Sven Eckelmann <sven.eckelmann@gmx.de> (author)
Signed-off-by: Simon Wunderlich <siwu@hrz.tu-chemnitz.de>
---
 batman/linux/modules/gateway.c   |    3 +++
 batman/linux/modules/gateway24.c |    3 +++
 2 files changed, 6 insertions(+), 0 deletions(-)



git-svn-id: http://downloads.open-mesh.org/svn/batman/trunk/batman@1109 45894c77-fb22-0410-b583-ff6e7d5dbf6c