Sync hash implementation of batgat and batman-adv
The compare functions had a different interpretation of its return value and
kmalloc can sleep inside batgat because we are running in user context of
kernel.

This patch can ignored if not wanted. It is just a copy of batman-adv-
kernelland/batman-core/hash.c to batman/linux/modules/hash.c so they have
nearly the same implementations. The only difference should now that the one
in batman-adv-kernelland uses GFP_ATOMIC in kmallocs and the one in batgat
uses GFP_KERNEL. Most changes are coding style changes and a different
interpretation of the return value of the compare functions.

Author: Sven Eckelmann <sven.eckelmann@gmx.de>


git-svn-id: http://downloads.open-mesh.org/svn/batman/trunk/batman@1120 45894c77-fb22-0410-b583-ff6e7d5dbf6c