Detect unstable version in SOURCE_VERSION
The SOURCE_VERSION consists of a version part and the actual status of
this version (called postfix). Both parts are seperated by a dash.
This seperator dash can be detected during runtime and can enable the
unstable warning instead of removing this warning in every release.
Before releasing the a new stable version the release manager must
remove the postfix of SOURCE_VERSION in batman.h, check INSTALL,
THANKS and README and then create the tag in the repository as usual.
Afterwards the development branch should change the SOURCE_VERSION to
the upcoming version number including a postfix starting with a dash
that informs the user that it is "unstable" and in development. This
could be for example "-dev", "-pre-alpha", "-beta" or "-rc1".

Signed-off-by: Sven Eckelmann <sven.eckelmann@gmx.de>


git-svn-id: http://downloads.open-mesh.org/svn/batman/trunk/batman@1178 45894c77-fb22-0410-b583-ff6e7d5dbf6c