Add compilation support for GNU/kFreeBSD
Debian recently announced that it added GNU/kFreeBSD as supported
distribution. It consists of an FreeBSD kernel and the usual debian
userspace tools and libraries. It uses the glibc with some extra
(free)bsd libraries to support bsd-only functions like devname and
strlcpy.
batmand should instantly work on this platform if a BSD developer would
port the bsd compatibility layer again to (Open|Free)BSD.

Signed-off-by: Sven Eckelmann <sven.eckelmann@gmx.de>


git-svn-id: http://downloads.open-mesh.org/svn/batman/trunk/batman@1244 45894c77-fb22-0410-b583-ff6e7d5dbf6c