[batman] Add padding around allocation	debugger structures
Architectures with a special alignment for load and store operations on
datatypes bigger than bytes will return a prealigned memory region when
calling malloc. When we add our data structure before and after this
region we destroy this alignment.
To fix this problem we add special regions with "magic" padding data.
To be sure that it is big enough for every load/store operation we use
the alignment for uintmax_t or a pointer even when the architecture only
supports smaller load/store operations.

Signed-off-by: Sven Eckelmann <sven.eckelmann@gmx.de>


git-svn-id: http://downloads.open-mesh.org/svn/batman/trunk/batman@1279 45894c77-fb22-0410-b583-ff6e7d5dbf6c